fix float to u64 intrinsics
reverts commit
310e3b428cfd181b51723276e6563b90d670da06
(more info there)

now functions check for
sign bit in float.

now hopefully this patch will
cover entirety of areas it might affect

+++ tccgen.c
@@ -3478,7 +3478,7 @@ again:
                 vtop->c.i = (vtop->c.ld != 0);
             } else {
                 if(sf)
-                    vtop->c.i = (int64_t)vtop->c.ld;
+                    vtop->c.i = vtop->c.ld;
                 else if (sbt_bt == VT_LLONG || (PTR_SIZE == 8 && sbt == VT_PTR))
                     ;
                 else if (sbt & VT_UNSIGNED)
